+++ backend/routes/api/admin/routes.ts
@@ -1,0 +1,7 @@
+import express from 'express'
+const router = express.Router()
+import userRoutes from './user.js'
+
+router.use('/user', userRoutes)
+
+export default router
