+++ backend/routes/api/approve/routes.ts
@@ -1,0 +1,7 @@
+import express from 'express'
+const router = express.Router()
+import travelRoutes from './travel.js'
+
+router.use('/travel', travelRoutes)
+
+export default router
