+++ backend/routes/api/approve/travel.ts
@@ -1,11 +1,11 @@
-import { getter, setter } from '../helper.js'
-import Travel, { TravelDoc } from '../models/travel.js'
-import { sendNotificationMail } from '../mail/mail.js'
+import { getter, setter } from '../../../helper.js'
+import Travel, { TravelDoc } from '../../../models/travel.js'
+import { sendNotificationMail } from '../../../mail/mail.js'
 import express, { Request, Response } from 'express'
-import { Travel as ITravel } from '../../common/types.js'
+import { Travel as ITravel } from '../../../../common/types.js'
 const router = express.Router()
 
-router.get('/travel', async (req, res) => {
+router.get('/', async (req, res) => {
   const sortFn = (a: ITravel, b: ITravel) => (a.startDate as Date).valueOf() - (b.startDate as Date).valueOf()
   return getter(
     Travel,
@@ -17,7 +17,7 @@ router.get('/travel', async (req, res) => {
   )(req, res)
 })
 
-router.get('/travel/approved', async (req, res) => {
+router.get('/approved', async (req, res) => {
   const sortFn = (a: ITravel, b: ITravel) => (b.startDate as Date).valueOf() - (a.startDate as Date).valueOf() // sort backwards
   return getter(
     Travel,
@@ -51,7 +51,7 @@ function approve(state: 'approved' | 'rejected') {
     return setter(Travel, '', false, check, sendNotificationMail)(req, res)
   }
 }
-router.post('/travel/approved', approve('approved'))
-router.post('/travel/rejected', approve('rejected'))
+router.post('/approved', approve('approved'))
+router.post('/rejected', approve('rejected'))
 
 export default router
